Move istio modifications to values.yaml.
We ran into repeated problems with the named modified istio directories not containing the changes their names claimed. Instead, we'll use the helm templating functionality to fill in the values changes we need.

In future commits, we expect to treat values.yaml similar to vars.sh -- i.e. it should be copied to each run's output directory and should be accessible from the results page.