Isolate Prometheus on its own node.
On the first 4k-workload runs that finished, Prometheus was evicted for using too much memory (23GB RAM).
If it has its own node, hopefully it will be allowed a bit more space.
If that is still not enough, we can increase node size to increase how much RAM prometheus can have.