Make route poller only use half the CPUs.
When it gets to use all 16 CPUs on the client, the machine is nearly fully utilized before we even start ending CP load. I'm worried that if we add this much load on the client side, we might end up sending config more slowly than desired and not loading Envoy as heavily on the CP side.

Signed-off-by: Mikael Manukyan <mmanukyan@pivotal.io>